Added `namespace` to Android build.gradle and removed irrelevant log messages

+++ lib/src/localstore/utils/io.dart
@@ -202,10 +202,12 @@ class Utils implements UtilsImpl {
   Future _deleteFile(String path) async {
     final dbDir = await Localstore.instance.databaseDirectory;
     final file = File('${dbDir.path}$path');
-    try {
-      await file.delete();
-    } catch (e) {
-      _log.finest(e);
+    if (await file.exists()) {
+      try {
+        await file.delete();
+      } catch (e) {
+        _log.finest(e);
+      }
     }
     _fileCache.remove(path);
   }
@@ -213,10 +215,12 @@ class Utils implements UtilsImpl {
   Future _deleteDirectory(String path) async {
     final dbDir = await Localstore.instance.databaseDirectory;
     final dir = Directory('${dbDir.path}$path');
-    try {
-      await dir.delete(recursive: true);
-    } catch (e) {
-      _log.finest(e);
+    if (await dir.exists()) {
+      try {
+        await dir.delete(recursive: true);
+      } catch (e) {
+        _log.finest(e);
+      }
     }
     _fileCache.removeWhere((key, value) => key.startsWith(path));
   }
